requirements.txt
Package              Version            
-------------------- -------------------
absl-py              0.9.0              
astor                0.8.1              
certifi              2019.11.28         
click                7.1.1              
Flask                1.1.1              
gast                 0.2.2              
gevent               1.4.0              
google-pasta         0.2.0              
greenlet             0.4.15             
grpcio               1.27.2             
h5py                 2.10.0             
itsdangerous         1.1.0              
Jinja2               2.11.1             
Keras-Applications   1.0.8              
Keras-Preprocessing  1.1.0              
Markdown             3.2.1              
MarkupSafe           1.1.1              
numpy                1.18.2             
opencv-python        4.2.0.32           
opt-einsum           3.2.0              
Pillow               7.0.0              
pip                  20.0.2             
protobuf             3.11.3             
pygame               1.9.6              
setuptools           46.0.0.post20200309
six                  1.14.0             
tensorboard          1.15.0             
tensorflow           1.15.0             
tensorflow-estimator 1.15.1             
termcolor            1.1.0              
Werkzeug             1.0.0              
wheel                0.34.2             
wrapt                1.12.1             
